test: AOP 미션 2단계
