+++ study/src/test/java/aop/stage1/Stage1Test.java
@@ -7,8 +7,6 @@ import aop.repository.UserDao;
 import aop.repository.UserHistoryDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.aop.framework.ProxyFactoryBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -19,8 +17,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class Stage1Test {
-
-    private static final Logger log = LoggerFactory.getLogger(Stage1Test.class);
 
     @Autowired
     private UserDao userDao;
