+++ study/src/test/java/aop/stage2/Stage2Test.java
@@ -5,8 +5,6 @@ import aop.StubUserHistoryDao;
 import aop.domain.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -16,11 +14,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class Stage2Test {
 
-    private static final Logger log = LoggerFactory.getLogger(Stage2Test.class);
-
     @Autowired
     private UserService userService;
-
     @Autowired
     private StubUserHistoryDao stubUserHistoryDao;
