+++ study/src/test/java/aop/stage2/UserService.java
@@ -1,42 +1,15 @@
 package aop.stage2;
 
-import aop.common.Transactional;
 import aop.domain.User;
-import aop.domain.UserHistory;
-import aop.repository.UserDao;
 import aop.repository.UserHistoryDao;
-import org.springframework.stereotype.Service;
 
-@Service
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private UserHistoryDao userHistoryDao;
+    User findById(long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
+    void insert(User user);
 
-    @Transactional
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
+    void changePassword(long id, String newPassword, String createBy);
 
-    @Transactional
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    @Transactional
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final var user = findById(id);
-        user.changePassword(newPassword);
-        userDao.update(user);
-        userHistoryDao.log(new UserHistory(user, createBy));
-    }
-
-    public void setUserHistoryDao(final UserHistoryDao userHistoryDao) {
-        this.userHistoryDao = userHistoryDao;
-    }
+    void setUserHistoryDao(UserHistoryDao userHistoryDao);
 }
